builtin/index-pack.c: separate reverse-index handling
To determine whether or not index-pack will write a .rev file, it checks
whether or not either of the WRITE_REV or WRITE_REV_VERIFY flags are set
in struct pack_idx_option.

This is confusing since we treat the presence of either to indicate
whether or not we want to either write or verify a .rev file, but
ultimately decide which based onw hether or not --verify was passed. In
other words, the combination of having the WRITE_REV bit set in flags
along with `--verify` translates to unsetting WRITE_REV and setting
WRITE_REV_VERIFY in its place.

This makes it overly confusing to switch the default, so let's
handle the decision of whether or not to write a reverse index by
updating a single variable and setting the flags accordingly.

Signed-off-by: Taylor Blau <me@ttaylorr.com>