pack-revindex: invert GIT_TEST to disable pack.writeReverseIndex
When developing the on-disk format for the pack reverse index, we used
the GIT_TEST_WRITE_REV_INDEX environment variable to control the default
value of pack.writeReverseIndex.

When GIT_TEST_WRITE_REV_INDEX=1, we would set pack.writeReverseIndex to
true, forcing the test suite to write and use the on-disk reverse index.

Now that pack.writeReverseIndex is true by default, change the meaning
of this variable by replacing it with GIT_TEST_IN_CORE_REV_INDEX. When
this new environment variable is set, do the opposite: avoid generating
an on-disk reverse index forcing the test suite to compute and use the
legacy in-core one instead.

There are some tests which ensure parity between the two (c.f., t5325.8
for an example), but coverage is limited (for e.g., that test does not
ever call pack_pos_to_index in the in-core version, reducing its
overall coverage). For full coverage, let's continue to run the test
suite in this alternative mode during CI to ensure we don't regress the
in-core reverse index code.

Signed-off-by: Taylor Blau <me@ttaylorr.com>