core.fsync: introduce granular fsync control
This commit introduces the `core.fsync` configuration
knob which can be used to control how components of the
repository are made durable on disk.

This setting allows future extensibility of the list of
syncable components:
* We issue a warning rather than an error for unrecognized
  components, so new configs can be used with old Git versions.
* We support negation, so users can choose one of the aggregate
  options and then remove components that they don't want.
  Aggregate options are defined in a later patch in this series.

This also supports the common request of doing absolutely no
fysncing with the `core.fsync=none` value, which is expected
to make the test suite faster.

Complete documentation for the new setting is included in a later patch
in the series so that it can be reviewed in final form.

Signed-off-by: Neeraj Singh <neerajsi@microsoft.com>