bisect: move even the command-line parsing to `bisect--helper`
On our journey to a fully built-in `git bisect`, this is the
last step.

Side note: The `parse-options` API is not at all set up to parse
subcommands such as `git bisect start`, `git bisect reset`, etc.
Instead of fighting an up-hill battle trying to "fix" that, we simply
roll the same type of manual subcommand parsing as we already do e.g.
in `builtin/bundle.c`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>