rebase --apply: improve fast-forward reflog message
Using move_to_original_branch() when reattaching HEAD after a
fast-forward improves HEAD's reflog message when rebasing a branch. This
is because it uses separate reflog messages for "HEAD" and
"branch". HEAD's reflog will now record which branch we're returning to.

When rebasing a detached HEAD there is no change in behavior. We
currently call reset_head() when head_name is NULL but
move_to_original_branch() does not. However the existing call does not
add a message to the reflog because we're not changing the commit that
HEAD points to and so lock_ref_for_update() skips the update.

Note that the removal of "strbuf_reset(&msg)" is safe as there is a call
to strbuf_release(&msf) just above this hunk which can be seen by
viewing the diff with -U6.

Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>