config: add git_config_get_timestamp()
The existing config parsing methods do not include a way to consistently
parse timestamps across all platforms. Recall that "unsigned long" is
32 bits on 64-bit Windows, so git_config_get_ulong() is insufficient.

Adding a new type requires quite a bit of boilerplate to match the style
of other types.

RFC-QUESTION: Would this be better to use uintmax_t, which could be cast
to timestamp_t or other types more robust than "unsigned long"?

Signed-off-by: Derrick Stolee <derrickstolee@github.com>