merge: new autosetupmerge option 'simple' for matching branches
The push.defaut option "simple" helps produce
predictable/understandable behavior for beginners,
where they don't accidentally push to the
"wrong" branch in centralized workflows. If they
create a local branch with a different name
and then try to do a plain push, it will
helpfully fail and explain why.

However, such users can often find themselves
confused by the behavior of git after they first
branch, and before they push. At that stage,
their upstream tracking branch is the original
remote branch, and pull (for example) behaves
very differently to how it later does when they
create their own same-name remote branch.

This commit introduces a new option to the
branch.autosetupmerge setting, "simple",
which is intended to be consistent with and
complementary to the push.default "simple"
option.

It will set up automatic tracking for a new
branch only if the remote ref is a branch and
that remote branch name matches the new local
branch name. It is a reduction in scope of
the existing default option, "true".

Signed-off-by: Tao Klerks <tao@klerks.biz>