rebase: note `preserve` merges may be a pull config option
The `--preserve-merges` option was removed by v2.35.0. However
users may not be aware that it is also a Pull option, and it is
still offered by major IDE vendors such as Visual Studio.

Extend the `--preserve-merges` die message to also direct users to
the use of the `preserve` option in the `pull` config.

Signed-off-by: Philip Oakley <philipoakley@iee.email>